backend/analysis_engines/drug_discovery.py
import torch
import torch.nn as nn
import torch.nn.functional as F
from torch_geometric.nn import GCNConv, global_mean_pool
try:
    from rdkit import Chem
    from rdkit.Chem import Descriptors
except ImportError:
    Chem = None
    Descriptors = None

class DrugDiscoveryGNN(nn.Module):
    """
    Graph Neural Network for predicting Drug-Protein interactions.
    Input: Heterogeneous graph of drugs and proteins.
    Output: Interaction probability.
    """
    def __init__(self, input_dim, hidden_dim, output_dim):
        super(DrugDiscoveryGNN, self).__init__()
        self.conv1 = GCNConv(input_dim, hidden_dim)
        self.conv2 = GCNConv(hidden_dim, hidden_dim)
        self.fc = nn.Linear(hidden_dim, output_dim)

    def forward(self, x, edge_index, batch):
        # 1. Obtain node embeddings 
        x = self.conv1(x, edge_index)
        x = x.relu()
        x = self.conv2(x, edge_index)
        x = x.relu()

        # 2. Readout layer (graph level representation)
        x = global_mean_pool(x, batch)

        # 3. Final classifier
        x = F.dropout(x, p=0.5, training=self.training)
        x = self.fc(x)
        
        return torch.sigmoid(x)

class DrugRepurposingEngine:
    def __init__(self, model_path=None):
        self.device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
        self.model = DrugDiscoveryGNN(input_dim=128, hidden_dim=64, output_dim=1).to(self.device)
        if model_path:
            self.model.load_state_dict(torch.load(model_path))
        self.model.eval()

    def predict(self, drug_smiles, protein_sequence):
        """
        Placeholder for predicting interaction between a drug and a protein.
        In a real scenario, SMILES and sequences are converted to graphs.
        """
        # 1. Logic to convert SMILES to Graph
        num_atoms = 10
        if Chem is not None:
            mol = Chem.MolFromSmiles(drug_smiles)
            if mol:
                num_atoms = mol.GetNumAtoms()
        
        # Mock node features and edges based on num_atoms for the GNN
        # input_dim is 128
        x = torch.rand((num_atoms, 128)).to(self.device)
        
        # Fully connected mock graph for edge_index
        rows, cols = [], []
        for i in range(num_atoms):
            for j in range(num_atoms):
                if i != j:
                    rows.append(i)
                    cols.append(j)
                    
        if len(rows) == 0:
            # Handle single atom edge case
            rows, cols = [0], [0]
            
        edge_index = torch.tensor([rows, cols], dtype=torch.long).to(self.device)
        batch = torch.zeros(num_atoms, dtype=torch.long).to(self.device)

        with torch.no_grad():
            output = self.model(x, edge_index, batch)
            mock_score = output.item()
            
        return mock_score

if __name__ == "__main__":
    engine = DrugRepurposingEngine()
    print("AI Analysis Engine Initialized.")
