backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
biopython
rdkit-pypi
scanpy
networkx
pandas
numpy<2
scikit-learn
torch
torch-geometric
httpx
python-dotenv
